+++ src/index.js
@@ -25,7 +25,7 @@ export default (config) => {
     privacy_url: '/privacy',
     ...(config.settings.DSGVOBanner || {}),
   };
-  config.settings.loadables['reactGa'] = loadable.lib(() => import('react-ga'));
+  config.settings.loadables['reactGa'] = loadable.lib(() => import('react-ga4'));
   config.settings.loadables['matomoTracker'] = loadable.lib(() =>
     import('@datapunt/matomo-tracker-js'),
   );