+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Feature
 
+- Upgrade react-ga4 in order to also support Google Analytics GA4 @avoinea refs #9
+
 ### Bugfix
 
 ### Internal